chore: move Metronome in a separate file
